requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
drf-spectacular>=0.27
django-cors-headers>=4.3
Pillow>=10.1
python-magic-bin>=0.4.14
django-filter>=23.5
qrcode>=7.4
